Second Minus Pipe Goes to World 5
In the Japanese re-release of SMB which is what this game's Minus World is based on, while the first and third pipes take you to the Minus world, the second pipe takes you to World 5.

This PR fixes issue #64 and warps the player to World 5 when going into the second pipe rather than the first.