Updated Red Coins and Coin Sparkle
Was making my rounds with tweaking the Levels with Level Maps from Mario Universe and TMK, but I made a detour to mess with the money.
This PR adds the unused Coin sparkle effects to regular coins, and adjusts the conditions of the Red Coin. I also expanded the `ResourceSetterNew` class to check for properties that are Objects which can be recursively edited with the identifier "."
Let me know if you prefer the sparkling to be off by default on either or both.

- Adjust ResourceSetterNew so it can also affect properties that are Objects (recursive property changing basically)
- Added Coin Sparkle on by default for Coins in Blocks or floating in the Level. Can be affected by Resource Packs and JSON changes
- When hitting a Coin Brick Block after some time, it produces 1 coin after hitting again instead of 2
- The Red Coin will only pop out of Coin Brick Blocks after being hit at least 9 times prior to the timeout. On the 10th hit or greater, the coin will pop out.