Door Pose Fix for Custom Levels
Fixes an issue with the DoorPose animation and detection I added where if there was another End Castle behind you, you'd get sent back in the level. This is primarily an issue for Custom Levels, and is extremely noticeable when reloading resource packs. And in the video below, it also broke the Door Pose animation in the level.

This PR adds an extra check by using a collision shape to make sure the player is in the right area before applying the positional changes.