Tweaked "Power-up Jump" to be more accurate
The Power-up Jump bug from the original game is supposed to be a trick that is relatively easy to pull off, rather than a frame perfect trick as it was coded before.

As long as the player isn't holding the jump button, they can hold the button through the power-up sequence, and as long as they are holding it, the player will perform the power-up jump.